feat: refactor project structure before starting on issue #13 . readwrite becomes parsingJava and project
