+++ src/parsingJava.js
@@ -1,207 +1,36 @@
 const fs = require('fs');
 const path = require('path');
 const { parse } = require('java-parser');
+const projectClasses = require('./project')
 
-class Project {
-    constructor(packages, classes) {
-        this.packages = packages || [];
-        this.classes = classes || [];
-        this.listOfClassNames = [];
-    }
-    addPackage(pkg) {
-        this.packages.push(pkg);
-    }
-    addClass(clazz) {
-        this.classes.push(clazz);
-    }
-    toUML() {
-        const uml = [];
-        var depth = 0;
-        uml.push('@startuml');
-        for (const pkg of this.packages) {
-            pkg.toUML(uml, depth);
-        }
-        for (const clazz of this.classes) {
-            clazz.toUML(uml, depth);
-        }
-        for (const clazz of this.classes) {
-            this.addContainArrows(uml, clazz);
-        }
-        for (const pkg of this.packages) {
-            this.addContainArrowsFromPackages(uml, pkg);
-        }
-        uml.push('@enduml');
-        return uml;
-    }
-    addContainArrows(uml, clazz) {
-        const dictOfContainedClasses = new dict();
-        for (const containes of clazz.containedClasses) {
-            if (this.listOfClassNames.includes(containes)) {
-                dictOfContainedClasses.add(containes);
-            }
-        }
-        for (const containedClass of dictOfContainedClasses.getAll()) {
-            if (dictOfContainedClasses.get(containedClass) == 1) {
-                uml.push(`  ${clazz.name} o-- ${containedClass}: contains`);
-            }else if (dictOfContainedClasses.get(containedClass) > 1) {
-                uml.push(`  ${clazz.name} "1" o-- "many" ${containedClass}: contains`);
-            }
-        }
-    }
-    addContainArrowsFromPackages(uml, pkg) {
-        for (const clazz of pkg.classes) {
-            this.addContainArrows(uml, clazz);
-        }
-        for (const subPackage of pkg.containedPackages) {
-            this.addContainArrowsFromPackages(uml, subPackage);
-        }
-    }
-}
-
-class dict {
-    constructor() {
-        this.data = {};
-        this.itterbleList = [];
-    }
-    add(key) {
-        if (!this.itterbleList.includes(key)) {
-            this.data[key] = 1;
-            this.itterbleList.push(key);
-        }else {
-            this.data[key]++;
-        }
-    }
-    get(key) {
-        return this.data[key] || 0;
-    }
-    getAll() {
-        return this.itterbleList;
-    }
-}
-
-class Package {
-    constructor(name, classes, containedPackages) {
-        this.name = name;
-        this.classes = classes || [];
-        this.containedPackages = containedPackages || [];
-    }
-    addClass(clazz) {
-        this.classes.push(clazz);
-    }
-    addPackage(pkg) {
-        this.containedPackages.push(pkg);
-    }
-    toUML(uml, depth) {
-        const indent = ' '.repeat(depth * 2);
-        uml.push(`${indent}package "${this.name}" {`);
-        for (const clazz of this.classes) {
-            clazz.toUML(uml, depth + 1);
-        }
-        for (const pkg of this.containedPackages) {
-            pkg.toUML(uml, depth + 1);
-        }
-        uml.push(`${indent}}`);
-    }
-}
-
-class Class {
-    constructor(name, methods, fields, superclass, implementedInterfaces, containedClasses, isAbstract = false, isInterface = false, isEnum = false) {
-        this.name = name;
-        this.methods = methods || [];
-        this.fields = fields || [];
-        this.superclass = superclass || null;
-        this.implementedInterfaces = implementedInterfaces || [];
-        this.containedClasses = containedClasses || [];
-        this.isAbstract = isAbstract;
-        this.isInterface = isInterface;
-        this.isEnum = isEnum;
-    }
-
-    addMethod(method) {
-        this.methods.push(method);
-    }
-
-    addField(field) {
-        this.fields.push(field);
-    }
-
-    addContainedClass(clazz) {
-        this.containedClasses.push(clazz);
-    }
-
-    toUML(uml, depth) {
-        const indent = ' '.repeat(depth * 2);
-        var type = 'class ';
-        if (this.isAbstract) {
-            type = 'abstract ';
-        }else if (this.isInterface) {
-            type = 'interface ';
-        }else if (this.isEnum) {
-            type = 'enum ';
-        }
-        let classDef = `${indent}${type}${this.name}`;
-        if (this.superclass) {
-            classDef += ` extends ${this.superclass}`;
-        }
-        if (this.implementedInterfaces.length > 0) {
-            classDef += ` implements ${this.implementedInterfaces.join(', ')}`;
-        }
-        uml.push(`${indent}${classDef} {`);
-        const innerIndent = ' '.repeat((depth + 1) * 2);
-        for (const field of this.fields) {
-            const visibility = field.isPrivate ? '-' : '+';
-            uml.push(`${innerIndent}  ${visibility} ${field.name}: ${field.type}`);
-        }
-        for (const method of this.methods) {
-            const visibility = method.isPrivate ? '-' : '+';
-            const params = method.parameters.map(p => `${p.name}: ${p.type}`).join(', ');
-            uml.push(`${innerIndent}  ${visibility} ${method.name}(${params}): ${method.returnType}`);
-        }
-        uml.push(`${indent}}`);
-    }
-}
-
-class Field {
-    constructor(name, type, isPrivate = false) {
-        this.name = name;
-        this.type = type;
-        this.isPrivate = isPrivate;
-    }
-}
-
-class Method {
-    constructor(name, returnType, parameters, isPrivate = false) {
-        this.name = name;
-        this.returnType = returnType;
-        this.parameters = parameters || [];
-        this.isPrivate = isPrivate;
-    }
-}
-
-class parameter {
-    constructor(name, type) {
-        this.name = name;
-        this.type = type;
-    }
-}
-
+//Converts Java to project object
 function readSourceFolder(sourceFolder) {
+
+    //init all objects
     const project = readProject(sourceFolder);
+
+    //condenses packages to one package with the name Java/main/com/logic for example
     fixAllNestedPackages(project);
+
     const listOfClassNames = project.classes.map(clazz => clazz.name);
     for (const packageObj of project.packages) {
         findClassNames(listOfClassNames, packageObj);
     }
+
+    //find any fields in classes that are objects of other classes
     for (const classObj of project.classes) {
         addContainedClass(listOfClassNames, classObj);
     }
     for (const packageObj of project.packages) {
         addContainedClassOfPackages(listOfClassNames, packageObj);
     }
+
     project.listOfClassNames = listOfClassNames;
+
     return project;
 }
 
+// Four recursive funtions to fix package names
 function fixAllNestedPackages(project){
     for(const pkg of project.packages){
         fixNestedSubPackages(pkg);
@@ -229,10 +58,11 @@ function recursiveNestedPackageFinder(name, packageObj){
         name += '/' +innerPackage.name;
         return recursiveNestedPackageFinder(name, innerPackage);
     }else{
-        return new Package(name, packageObj.classes, packageObj.containedPackages);
+        return new projectClasses.Package(name, packageObj.classes, packageObj.containedPackages);
     }
 }
 
+//three funtions to find contained classes
 function addContainedClassOfPackages(listOfClassNames, packageObj) {
     for (const clazz of packageObj.classes) {
         addContainedClass(listOfClassNames, clazz);
@@ -261,13 +91,14 @@ function addContainedClass(listOfClassNames, classObj) {
     }
 }
 
+//recursive functions to read project
 function readProject(sourceFolder) {
 
     if (!fs.existsSync(sourceFolder)) {
         throw new Error(`Source folder does not exist: ${sourceFolder}`);
     }
 
-    const project = new Project();
+    const project = new projectClasses.Project();
 
     const classes = getClassesInPath(sourceFolder);
 
@@ -305,19 +136,16 @@ function getPackagesInPath(currentPath) {
         const stat = fs.statSync(fullPath);
         if (stat.isDirectory()) {
             const packageName = path.basename(fullPath);
-            // Collect classes in this directory
             const packageClasses = getClassesInPath(fullPath);
-            // Recursively collect contained packages
             const containedPackages = getPackagesInPath(fullPath);
-            // Create the package with its classes and contained packages
-            const packageObj = new Package(packageName, packageClasses, containedPackages);
+            const packageObj = new projectClasses.Package(packageName, packageClasses, containedPackages);
             packages.push(packageObj);
         }
     }
     return packages;
 }
 
-// Helper to recursively find the first .image property in a CST node
+//helper to recursively find the first .image property in a CST node
 function findTypeImage(node) {
     if (!node) return null;
     if (Array.isArray(node)) {
@@ -337,6 +165,9 @@ function findTypeImage(node) {
     return null;
 }
 
+
+//tbh this funtion is an absolut mess... should prolly tighty it up at some point
+//TODO: make ugly funtion pretty
 function getClassesInFile(filePath) {
     const classes = [];
     const code = fs.readFileSync(filePath, 'utf8');
@@ -344,10 +175,10 @@ function getClassesInFile(filePath) {
     try {
         cst = parse(code);
     } catch (e) {
-        // If parse fails, skip this file
+        //if parse fails skip this file
         return classes;
     }
-    // Find the root node (handle both CST shapes)
+    //find the root node (handle both CST shapes)
     let root = cst;
     if (root.name === 'compilationUnit' && root.children && root.children.ordinaryCompilationUnit) {
         root = root.children.ordinaryCompilationUnit[0];
@@ -362,24 +193,24 @@ function getClassesInFile(filePath) {
         let isAbstract = false;
         let name = 'Unknown';
         let classDecl = null;
-        // Check for enum
+        //check for enum
         if (typeDecl.children && typeDecl.children.enumDeclaration) {
             isEnum = true;
             classDecl = typeDecl.children.enumDeclaration[0];
-            // Enum name
+            //enum name
             if (classDecl.children.Identifier) {
                 name = classDecl.children.Identifier[0].image;
             }
         } else if (typeDecl.children && typeDecl.children.interfaceDeclaration) {
             isInterface = true;
             classDecl = typeDecl.children.interfaceDeclaration[0];
-            // Interface name
+            //interface name
             if (classDecl.children.normalInterfaceDeclaration && classDecl.children.normalInterfaceDeclaration[0].children.typeIdentifier) {
                 name = classDecl.children.normalInterfaceDeclaration[0].children.typeIdentifier[0].children.Identifier[0].image;
             }
         } else if (typeDecl.children && typeDecl.children.classDeclaration) {
             classDecl = typeDecl.children.classDeclaration[0];
-            // Abstract class detection
+            //abstract class detection
             if (classDecl.children.classModifier) {
                 for (const modifier of classDecl.children.classModifier) {
                     if (modifier.children && modifier.children.Abstract) {
@@ -388,7 +219,7 @@ function getClassesInFile(filePath) {
                     }
                 }
             }
-            // Class name
+            //class name
             if (
                 classDecl.children.normalClassDeclaration &&
                 classDecl.children.normalClassDeclaration[0].children.typeIdentifier &&
@@ -401,10 +232,10 @@ function getClassesInFile(filePath) {
         }
         let superclass = null;
         let implementedInterfaces = [];
-        // Only for normal classes (not enums/interfaces)
+        //only for normal classes (not enums/interfaces)
         if (!isEnum && !isInterface && classDecl.children.normalClassDeclaration) {
             const normalClass = classDecl.children.normalClassDeclaration[0];
-            // Superclass (support both 'superclass' and 'classExtends' CST shapes)
+            //superclass (support both 'superclass' and 'classExtends' CST shapes)
             if (normalClass.children.classExtends && normalClass.children.classExtends[0].children.classType) {
                 const classTypeNode = normalClass.children.classExtends[0].children.classType[0];
                 if (classTypeNode.children.Identifier) {
@@ -418,7 +249,7 @@ function getClassesInFile(filePath) {
                     superclass = idents.join('.') || null;
                 }
             }
-            // Implemented interfaces (support both 'classImplements' and 'superinterfaces' CST shapes)
+            //implemented interfaces (support both 'classImplements' and 'superinterfaces' CST shapes)
             if (normalClass.children.classImplements && normalClass.children.classImplements[0].children.interfaceTypeList) {
                 const interfaceList = normalClass.children.classImplements[0].children.interfaceTypeList[0];
                 if (interfaceList.children.interfaceType) {
@@ -444,8 +275,8 @@ function getClassesInFile(filePath) {
                 }
             }
         }
-        const clazz = new Class(name, [], [], superclass, implementedInterfaces, [], isAbstract, isInterface, isEnum);
-        // Find classBody
+        const clazz = new projectClasses.Class(name, [], [], superclass, implementedInterfaces, [], isAbstract, isInterface, isEnum);
+        //find classBody
         let classBody = null;
         if (
             classDecl.children.normalClassDeclaration &&
@@ -458,15 +289,15 @@ function getClassesInFile(filePath) {
             for (const bodyDecl of bodyDecls) {
                 if (!bodyDecl.children || !bodyDecl.children.classMemberDeclaration) continue;
                 const memberDecl = bodyDecl.children.classMemberDeclaration[0];
-                // Fields
+                //fields
                 if (memberDecl.children.fieldDeclaration) {
                     for (const fieldDecl of memberDecl.children.fieldDeclaration) {
-                        // Type
+                        //type
                         let type = 'Object';
                         if (fieldDecl.children.unannType && fieldDecl.children.unannType[0]) {
                             type = findTypeImage(fieldDecl.children.unannType[0]) || 'Object';
                         }
-                        // Check for private modifier
+                        //check for private modifier
                         let isPrivate = false;
                         if (fieldDecl.children.fieldModifier) {
                             for (const modifier of fieldDecl.children.fieldModifier) {
@@ -476,45 +307,45 @@ function getClassesInFile(filePath) {
                                 }
                             }
                         }
-                        // Variable declarators
+                        //variable declarators
                         if (fieldDecl.children.variableDeclaratorList) {
                             const varDecls = fieldDecl.children.variableDeclaratorList[0].children.variableDeclarator;
                             for (const varDecl of varDecls) {
                                 const fieldName = varDecl.children.variableDeclaratorId[0].children.Identifier[0].image;
-                                clazz.addField(new Field(fieldName, type, isPrivate));
+                                clazz.addField(new projectClasses.Field(fieldName, type, isPrivate));
                             }
                         }
                     }
                 }
-                // Methods
+                //methods
                 if (memberDecl.children.methodDeclaration) {
                     for (const methodDecl of memberDecl.children.methodDeclaration) {
                         const methodHeader = methodDecl.children.methodHeader[0];
                         const methodName = methodHeader.children.methodDeclarator[0].children.Identifier[0].image;
-                        // Return type
+                        //return type
                         let returnType = 'void';
                         if (methodHeader.children.result && methodHeader.children.result[0].children.unannType) {
                             returnType = findTypeImage(methodHeader.children.result[0].children.unannType[0]) || 'void';
                         }
-                        // Check for private modifier (robust)
+                        //check for private modifier (robust)
                         let isPrivate = false;
                         if (methodDecl.children.methodModifier) {
                             for (const modifier of methodDecl.children.methodModifier) {
-                                // Each modifier may have multiple children (e.g., Private, Static, etc.)
+                                //each modifier may have multiple children (e.g., Private, Static, etc.)
                                 if (modifier.children && modifier.children.Private) {
                                     isPrivate = true;
                                     break;
                                 }
                             }
                         }
-                        // Use the parameter class for method parameters
+                        //use the parameter class for method parameters
                         let parameters = [];
                         if (methodHeader.children.methodDeclarator[0].children.formalParameterList) {
                             const paramList = methodHeader.children.methodDeclarator[0].children.formalParameterList[0];
-                            // Only handle formalParameter (not varargs for now)
+                            //only handle formalParameter (not varargs for now)
                             if (paramList.children.formalParameter) {
                                 for (const param of paramList.children.formalParameter) {
-                                    // Most common case: variableParaRegularParameter
+                                    //most common case: variableParaRegularParameter
                                     if (param.children.variableParaRegularParameter) {
                                         const regParam = param.children.variableParaRegularParameter[0];
                                         let paramType = 'Object';
@@ -525,21 +356,20 @@ function getClassesInFile(filePath) {
                                         if (regParam.children.variableDeclaratorId && regParam.children.variableDeclaratorId[0].children.Identifier) {
                                             paramName = regParam.children.variableDeclaratorId[0].children.Identifier[0].image;
                                         }
-                                        parameters.push(new parameter(paramName, paramType));
+                                        parameters.push(new projectClasses.Parameter(paramName, paramType));
                                     } else if (param.children.unannType && param.children.variableDeclaratorId) {
-                                        // Fallback for other parameter shapes
+                                        //fallback for other parameter shapes
                                         let paramType = findTypeImage(param.children.unannType[0]) || 'Object';
                                         let paramName = 'unknown';
                                         if (param.children.variableDeclaratorId[0].children.Identifier) {
                                             paramName = param.children.variableDeclaratorId[0].children.Identifier[0].image;
                                         }
-                                        parameters.push(new parameter(paramName, paramType));
+                                        parameters.push(new projectClasses.Parameter(paramName, paramType));
                                     }
                                 }
                             }
-                            // TODO: handle varargs (lastFormalParameter) if needed
                         }
-                        clazz.addMethod(new Method(methodName, returnType, parameters, isPrivate));
+                        clazz.addMethod(new projectClasses.Method(methodName, returnType, parameters, isPrivate));
                     }
                 }
             }
@@ -549,6 +379,7 @@ function getClassesInFile(filePath) {
     return classes;
 }
 
+//uses the project toUML() to write plantUML
 function writeUMLToFile(uml, outputLocation){
     if (!fs.existsSync(outputLocation)) {
         fs.mkdirSync(outputLocation, { recursive: true });
